+++ Cargo.toml
@@ -64,7 +64,7 @@ aho-corasick = "1.1"
 anyhow = "1.0"
 async-compression = { version = "0.4", default-features = false, optional = true, features = ["brotli", "deflate", "gzip", "zstd", "tokio"] }
 bcrypt = { version = "0.15", optional = true }
-bytes = "1.7"
+bytes = "1.8"
 chrono = { version = "0.4", default-features = false, features = ["std", "clock"], optional = true }
 clap = { version = "4.5", features = ["derive", "env"] }
 clap_allgen = "0.2.0"
@@ -110,7 +110,7 @@ prometheus = { version = "0.13", optional = true }
 windows-service = "0.7"
 
 [dev-dependencies]
-bytes = "1.7"
+bytes = "1.8"
 serde_json = "1.0"
 
 [build-dependencies]
