+++ src/handler.rs
@@ -333,57 +333,3 @@ impl RequestHandler {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use headers::HeaderValue;
-    use hyper::{Method, Request};
-    use std::net::SocketAddr;
-
-    use crate::http_ext::MethodExt;
-    use crate::testing::fixtures::{fixture_req_handler, fixture_settings, REMOTE_ADDR};
-
-    #[tokio::test]
-    async fn check_allowed_methods() {
-        let settings = fixture_settings("toml/handler.toml");
-        let req_handler = fixture_req_handler(settings.general, settings.advanced);
-        let remote_addr = Some(REMOTE_ADDR.parse::<SocketAddr>().unwrap());
-
-        let methods = [
-            Method::CONNECT,
-            Method::DELETE,
-            Method::GET,
-            Method::HEAD,
-            Method::PATCH,
-            Method::POST,
-            Method::PUT,
-            Method::TRACE,
-        ];
-        for method in methods {
-            let mut req = Request::default();
-            *req.method_mut() = method.clone();
-            *req.uri_mut() = "http://localhost/assets/index.html".parse().unwrap();
-
-            match req_handler.handle(&mut req, remote_addr).await {
-                Ok(resp) => {
-                    if method.is_allowed() {
-                        assert_eq!(resp.status(), 200);
-                        assert_eq!(
-                            resp.headers().get("content-type"),
-                            Some(&HeaderValue::from_static("text/html"))
-                        );
-                        assert_eq!(
-                            resp.headers().get("server"),
-                            Some(&HeaderValue::from_static("Static Web Server"))
-                        );
-                    } else {
-                        assert_eq!(resp.status(), 405);
-                    }
-                }
-                Err(err) => {
-                    panic!("unexpected error: {err}")
-                }
-            };
-        }
-    }
-}
